add basic routing test:
